+++ assignment-02/java/src/main/java/assignment/controller/MainController.java
@@ -6,7 +6,7 @@ public interface MainController {
      * Updates the currentWaterLevel
      * @param waterLevel
      */
-    void addWaterLevelRecord(int waterLevel);
+    void addWaterLevelRecord(float waterLevel);
 
     /**
      * Changes the COM port
